server/requirements.txt
python-socketio==5.7.1
uvicorn==0.18.3
websockets==10.3